keyring::credential

Trait CredentialApi

pub trait CredentialApi {
    // Required methods
    fn set_secret(&self, password: &[u8]) -> Result<()>;
    fn get_secret(&self) -> Result<Vec<u8>>;
    fn delete_credential(&self) -> Result<()>;
    fn as_any(&self) -> &dyn Any;

    // Provided methods
    fn set_password(&self, password: &str) -> Result<()> { ... }
    fn get_password(&self) -> Result<String> { ... }
    fn get_attributes(&self) -> Result<HashMap<String, String>> { ... }
    fn update_attributes(&self, _: &HashMap<&str, &str>) -> Result<()> { ... }
    fn debug_fmt(&self, f: &mut Formatter<'_>) -> Result { ... }
}

The API that credentials implement.

Required Methods

fn set_secret(&self, password: &[u8]) -> Result<()>

Set the credential’s secret (a byte array).

This will persist the secret in the underlying store.

fn get_secret(&self) -> Result<Vec<u8>>

Retrieve a secret (a byte array) from the credential.

This has no effect on the underlying store. If there is no credential for this entry, a NoEntry error is returned.

fn delete_credential(&self) -> Result<()>

Delete the underlying credential, if there is one.

This is not idempotent if the credential existed! A second call to delete_credential will return a NoEntry error.

fn as_any(&self) -> &dyn Any

Return the underlying concrete object cast to Any.

This allows clients to downcast the credential to its concrete type so they can do platform-specific things with it (e.g., query its attributes in the underlying store).

Provided Methods

fn set_password(&self, password: &str) -> Result<()>

Set the credential’s password (a string).

This will persist the password in the underlying store.

fn get_password(&self) -> Result<String>

Retrieve the password (a string) from the underlying credential.

This has no effect on the underlying store. If there is no credential for this entry, a NoEntry error is returned.

fn get_attributes(&self) -> Result<HashMap<String, String>>

Get the secure store attributes on this entry’s credential.

Each credential store may support reading and updating different named attributes; see the documentation on each of the stores for details. Note that the keyring itself uses some of these attributes to map entries to their underlying credential; these controlled attributes are not available for reading or updating.

We provide a default (no-op) implementation of this method for backward compatibility with stores that don’t implement it.

fn update_attributes(&self, _: &HashMap<&str, &str>) -> Result<()>

Update the secure store attributes on this entry’s credential.

Each credential store may support reading and updating different named attributes; see the documentation on each of the stores for details. The implementation will ignore any attribute names that you supply that are not available for update. Because the names used by the different stores tend to be distinct, you can write cross-platform code that will work correctly on each platform.

We provide a default no-op implementation of this method for backward compatibility with stores that don’t implement it.

fn debug_fmt(&self, f: &mut Formatter<'_>) -> Result

The Debug trait call for the object.

This is used to implement the Debug trait on this type; it allows generic code to provide debug printing as provided by the underlying concrete object.

We provide a (useless) default implementation for backward compatibility with existing implementors who may have not implemented the Debug trait for their credential objects