keyring::mock

Struct MockCredential

pub struct MockCredential {
    pub inner: Mutex<RefCell<MockData>>,
}

The concrete mock credential

Mocks use an internal mutability pattern since entries are read-only. The mutex is used to make sure these are Sync.

Fields
inner: Mutex<RefCell<MockData>>
Implementations

impl MockCredential

pub fn set_error(&self, err: Error)

Set an error to be returned from this mock credential.

Error returns always take precedence over the normal behavior of the mock. But once an error has been returned it is removed, so the mock works thereafter.